Ask groff for text instead of postscript on Solaris
In the last case fallback to groff in #ifdef HANDLE_ROFFSEQ, it
was just running "groff -man", which generates postscript output,
which xman displays as a text dump of the postscript source, not
very readable.

Hit on Solaris since it wasn't caught by the earlier checks for
CSRG_BASED, linux, __CYGWIN__, and __DARWIN__, which cover most
other cases.

Signed-off-by: Alan Coopersmith <alan.coopersmith@oracle.com>